@@ -6,11 +6,7 @@ This program selects the graph storage type to use based on the LIGHTRAG_GRAPH_S
 and tests its basic and advanced operations.
 
 Supported graph storage types include:
-- NetworkXStorage
-- Neo4JStorage
-- MongoDBStorage
 - PGGraphStorage
-- MemgraphStorage
 """
 
 import asyncio
@@ -68,7 +64,7 @@ async def initialize_graph_storage():
     Returns the initialized storage instance.
     """
     # Get the graph storage type from environment variables
-    graph_storage_type = os.getenv('LIGHTRAG_GRAPH_STORAGE', 'NetworkXStorage')
+    graph_storage_type = os.getenv('LIGHTRAG_GRAPH_STORAGE', 'PGGraphStorage')
 
     # Validate the storage type
     try:
