@@ -12,7 +12,7 @@ Comprehensive test suite covering workspace isolation in LightRAG:
 7. Error Handling - Invalid workspace configurations
 8. Update Flags Workspace Isolation - Update flags properly isolated
 9. Empty Workspace Standardization - Empty workspace handling
-10. JsonKVStorage Workspace Isolation - Integration test for KV storage
+10. (removed) Legacy JsonKVStorage workspace isolation
 11. LightRAG End-to-End Workspace Isolation - Complete E2E test with two instances
 
 Total: 11 test scenarios
@@ -715,177 +715,9 @@ async def test_empty_workspace_standardization():
 
 
 # =============================================================================
-# Test 10: JsonKVStorage Workspace Isolation (Integration Test)
+# # =============================================================================
+# Test 10 removed (legacy JsonKVStorage)
 # =============================================================================
-
-
-@pytest.mark.offline
-@pytest.mark.asyncio
-async def test_json_kv_storage_workspace_isolation(keep_test_artifacts):
-    """
-    Integration test: Verify JsonKVStorage properly isolates data between workspaces.
-    Creates two JsonKVStorage instances with different workspaces, writes different data,
-    and verifies they don't mix.
-    """
-    # Purpose: Ensure JsonKVStorage respects workspace-specific directories and data.
-    # Scope: storage initialization, upsert/get_by_id operations, and filesystem layout
-    # inside the temporary working directory.
-    print('\n' + '=' * 60)
-    print('TEST 10: JsonKVStorage Workspace Isolation (Integration)')
-    print('=' * 60)
-
-    # Create temporary test directory under project temp/
-    test_dir = str(Path(__file__).parent.parent / 'temp/test_json_kv_storage_workspace_isolation')
-    if os.path.exists(test_dir):
-        shutil.rmtree(test_dir)
-    os.makedirs(test_dir, exist_ok=True)
-    print(f'\n   Using test directory: {test_dir}')
-
-    try:
-        initialize_share_data()
-
-        # Mock embedding function
-        async def mock_embedding_func(texts: list[str]) -> np.ndarray:
-            return np.random.rand(len(texts), 384)  # 384-dimensional vectors
-
-        # Global config
-        global_config = {
-            'working_dir': test_dir,
-            'embedding_batch_num': 10,
-        }
-
-        # Test 10.1: Create two JsonKVStorage instances with different workspaces
-        print('\nTest 10.1: Create two JsonKVStorage instances with different workspaces')
-
-        from lightrag.kg.json_kv_impl import JsonKVStorage
-
-        storage1 = JsonKVStorage(
-            namespace='entities',
-            workspace='workspace1',
-            global_config=global_config,
-            embedding_func=EmbeddingFunc(embedding_dim=384, max_token_size=8192, func=mock_embedding_func),
-        )
-
-        storage2 = JsonKVStorage(
-            namespace='entities',
-            workspace='workspace2',
-            global_config=global_config,
-            embedding_func=EmbeddingFunc(embedding_dim=384, max_token_size=8192, func=mock_embedding_func),
-        )
-
-        # Initialize both storages
-        await storage1.initialize()
-        await storage2.initialize()
-
-        print('   Storage1 created: workspace=workspace1, namespace=entities')
-        print('   Storage2 created: workspace=workspace2, namespace=entities')
-
-        # Test 10.2: Write different data to each storage
-        print('\nTest 10.2: Write different data to each storage')
-
-        # Write to storage1 (upsert expects dict[str, dict])
-        await storage1.upsert(
-            {
-                'entity1': {
-                    'content': 'Data from workspace1 - AI Research',
-                    'type': 'entity',
-                },
-                'entity2': {
-                    'content': 'Data from workspace1 - Machine Learning',
-                    'type': 'entity',
-                },
-            }
-        )
-        print('   Written to storage1: entity1, entity2')
-        # Persist data to disk
-        await storage1.index_done_callback()
-        print('   Persisted storage1 data to disk')
-
-        # Write to storage2
-        await storage2.upsert(
-            {
-                'entity1': {
-                    'content': 'Data from workspace2 - Deep Learning',
-                    'type': 'entity',
-                },
-                'entity2': {
-                    'content': 'Data from workspace2 - Neural Networks',
-                    'type': 'entity',
-                },
-            }
-        )
-        print('   Written to storage2: entity1, entity2')
-        # Persist data to disk
-        await storage2.index_done_callback()
-        print('   Persisted storage2 data to disk')
-
-        # Test 10.3: Read data from each storage and verify isolation
-        print('\nTest 10.3: Read data and verify isolation')
-
-        # Read from storage1
-        result1_entity1 = await storage1.get_by_id('entity1')
-        result1_entity2 = await storage1.get_by_id('entity2')
-
-        # Read from storage2
-        result2_entity1 = await storage2.get_by_id('entity1')
-        result2_entity2 = await storage2.get_by_id('entity2')
-
-        print(f'   Storage1 entity1: {result1_entity1}')
-        print(f'   Storage1 entity2: {result1_entity2}')
-        print(f'   Storage2 entity1: {result2_entity1}')
-        print(f'   Storage2 entity2: {result2_entity2}')
-
-        # Verify isolation (get_by_id returns dict)
-        assert result1_entity1 is not None, 'Storage1 entity1 should not be None'
-        assert result1_entity2 is not None, 'Storage1 entity2 should not be None'
-        assert result2_entity1 is not None, 'Storage2 entity1 should not be None'
-        assert result2_entity2 is not None, 'Storage2 entity2 should not be None'
-        assert result1_entity1.get('content') == 'Data from workspace1 - AI Research', (
-            'Storage1 entity1 content mismatch'
-        )
-        assert result1_entity2.get('content') == 'Data from workspace1 - Machine Learning', (
-            'Storage1 entity2 content mismatch'
-        )
-        assert result2_entity1.get('content') == 'Data from workspace2 - Deep Learning', (
-            'Storage2 entity1 content mismatch'
-        )
-        assert result2_entity2.get('content') == 'Data from workspace2 - Neural Networks', (
-            'Storage2 entity2 content mismatch'
-        )
-        assert result1_entity1.get('content') != result2_entity1.get('content'), (
-            'Storage1 and Storage2 entity1 should have different content'
-        )
-        assert result1_entity2.get('content') != result2_entity2.get('content'), (
-            'Storage1 and Storage2 entity2 should have different content'
-        )
-
-        print('✅ PASSED: JsonKVStorage - Data Isolation')
-        print('   Two storage instances correctly isolated: ws1 and ws2 have different data')
-
-        # Test 10.4: Verify file structure
-        print('\nTest 10.4: Verify file structure')
-        ws1_dir = Path(test_dir) / 'workspace1'
-        ws2_dir = Path(test_dir) / 'workspace2'
-
-        ws1_exists = ws1_dir.exists()
-        ws2_exists = ws2_dir.exists()
-
-        print(f'   workspace1 directory exists: {ws1_exists}')
-        print(f'   workspace2 directory exists: {ws2_exists}')
-
-        assert ws1_exists, 'workspace1 directory should exist'
-        assert ws2_exists, 'workspace2 directory should exist'
-
-        print('✅ PASSED: JsonKVStorage - File Structure')
-        print(f'   Workspace directories correctly created: {ws1_dir} and {ws2_dir}')
-
-    finally:
-        # Cleanup test directory (unless keep_test_artifacts is set)
-        if os.path.exists(test_dir) and not keep_test_artifacts:
-            shutil.rmtree(test_dir)
-            print(f'\n   Cleaned up test directory: {test_dir}')
-        elif keep_test_artifacts:
-            print(f'\n   Kept test directory for inspection: {test_dir}')
 
 
 # =============================================================================
